refactor: 컴포넌트 분리 및 스토리북 작성
