+++ src/components/PasswordForm/index.jsx
@@ -1,21 +1,60 @@
+import { useRef } from 'react'
 import PropTypes from 'prop-types'
-import InputBox from 'components/InputBox'
-import { FormWrapper, Label } from 'components/Form/style'
-import { PasswordWrapper, Dot } from 'components/PasswordForm/style'
-import { Fragment } from 'react'
 
-function PasswordForm({ label, inputInfo, size, onChange }) {
+import InputBox from 'components/common/InputBox'
+import { isInvalidPassword } from 'validation'
+
+import { FormWrapper, Label } from 'components/common/Form/style'
+import { PasswordWrapper, Dot } from 'components/PasswordForm/style'
+
+function PasswordForm({ password, setPassword }) {
+  const secondPasswordInputRef = useRef()
+
+  const handlePasswordChange = ({ target: { value } }, key) => {
+    if (isInvalidPassword(value)) return
+
+    setPassword((prev) => {
+      const newState = { ...prev }
+      newState[key] = value
+      return newState
+    })
+
+    if (value.length >= 1) {
+      secondPasswordInputRef.current.focus()
+    }
+  }
+
   return (
     <FormWrapper>
       <Label>
-        <label>{label}</label>
+        <label>카드 비밀번호</label>
       </Label>
       <PasswordWrapper>
-        {inputInfo.map((info) => (
-          <Fragment key={info.key}>
-            <InputBox inputInfo={[info]} size={size} onChange={onChange} />
-          </Fragment>
-        ))}
+        <InputBox
+          inputInfo={[
+            {
+              type: 'password',
+              id: 'first-password',
+              value: password.firstPassword,
+              key: 'firstPassword',
+            },
+          ]}
+          size={12}
+          onChange={handlePasswordChange}
+        />
+        <InputBox
+          inputInfo={[
+            {
+              type: 'password',
+              id: 'second-password',
+              value: password.secondPassword,
+              key: 'secondPassword',
+              ref: secondPasswordInputRef,
+            },
+          ]}
+          size={12}
+          onChange={handlePasswordChange}
+        />
         <Dot>•</Dot>
         <Dot>•</Dot>
       </PasswordWrapper>
@@ -23,16 +62,16 @@ function PasswordForm({ label, inputInfo, size, onChange }) {
   )
 }
 
-PasswordForm.propTypes = {
-  /**
-   * form label 문구
-   */
-  label: PropTypes.string,
-  inputInfo: PropTypes.array,
-  /**
-   * input 넓이
-   */
-  size: PropTypes.number,
-  onChange: PropTypes.func,
-}
+// PasswordForm.propTypes = {
+//   /**
+//    * form label 문구
+//    */
+//   label: PropTypes.string,
+//   inputInfo: PropTypes.array,
+//   /**
+//    * input 넓이
+//    */
+//   size: PropTypes.number,
+//   onChange: PropTypes.func,
+// }
 export default PasswordForm
