+++ src/validation/index.js
@@ -2,24 +2,24 @@ import { CARD_NUMBER, DUE_DATE, OWNER, CVC, PASSWORD } from 'constant'
 
 const isNumber = (value) => isNaN(value)
 
-const isInValidCardNumber = (value) =>
+const isInvalidCardNumber = (value) =>
   value.length > CARD_NUMBER.UNIT_LENGTH || isNumber(value)
 
-const isInValidDueDate = (value) =>
+const isInvalidDueDate = (value) =>
   value.length > DUE_DATE.UNIT_LENGTH || isNumber(value)
 
-const isInValidOwner = (value) => value.length > OWNER.MAX_LENGTH
+const isInvalidOwner = (value) => value.length > OWNER.MAX_LENGTH
 
-const isInValidCVC = (value) =>
+const isInvalidCVC = (value) =>
   value.length > CVC.UNIT_LENGTH || isNumber(value)
 
-const isInValidPassword = (value) =>
+const isInvalidPassword = (value) =>
   value.length > PASSWORD.UNIT_LENGTH || isNumber(value)
 
 export {
-  isInValidCardNumber,
-  isInValidDueDate,
-  isInValidOwner,
-  isInValidCVC,
-  isInValidPassword,
+  isInvalidCardNumber,
+  isInvalidDueDate,
+  isInvalidOwner,
+  isInvalidCVC,
+  isInvalidPassword,
 }
